cloud_storage: move definitions of segment state structs
These will no longer be internal to remote_partition, once
we start managing the population of materialized_segment_state
objects (and their associated readers) from a central place
to improve resource management.
