+++ src/v/cloud_storage/remote_partition.h
@@ -15,6 +15,7 @@
 #include "cloud_storage/partition_probe.h"
 #include "cloud_storage/remote.h"
 #include "cloud_storage/remote_segment.h"
+#include "cloud_storage/segment_state.h"
 #include "cloud_storage/types.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
@@ -42,6 +43,7 @@ namespace cloud_storage {
 using namespace std::chrono_literals;
 
 class partition_record_batch_reader_impl;
+struct materialized_segment_state;
 
 namespace details {
 
@@ -244,61 +246,7 @@ private:
     void gc_stale_materialized_segments(bool force_collection);
 
     friend struct offloaded_segment_state;
-
-    struct materialized_segment_state;
-
-    /// State that have to be materialized before use
-    struct offloaded_segment_state {
-        explicit offloaded_segment_state(model::offset bo);
-
-        std::unique_ptr<materialized_segment_state>
-        materialize(remote_partition& p, kafka::offset offset_key);
-
-        ss::future<> stop();
-
-        offloaded_segment_state offload(remote_partition*);
-
-        model::offset base_rp_offset;
-
-        offloaded_segment_state* operator->() { return this; }
-
-        const offloaded_segment_state* operator->() const { return this; }
-    };
-
-    /// State with materialized segment and cached reader
-    ///
-    /// The object represent the state in which there is(or was) at
-    /// least one active reader that consumes data from the
-    /// remote segment.
-    struct materialized_segment_state {
-        materialized_segment_state(
-          model::offset bo, kafka::offset offk, remote_partition& p);
-
-        void return_reader(std::unique_ptr<remote_segment_batch_reader> reader);
-
-        /// Borrow reader or make a new one.
-        /// In either case return a reader.
-        std::unique_ptr<remote_segment_batch_reader> borrow_reader(
-          const storage::log_reader_config& cfg,
-          retry_chain_logger& ctxlog,
-          partition_probe& probe);
-
-        ss::future<> stop();
-
-        offloaded_segment_state offload(remote_partition* partition);
-
-        /// Base offsetof the segment
-        model::offset base_rp_offset;
-        /// Key of the segment in _segments collection of the remote_partition
-        kafka::offset offset_key;
-        ss::lw_shared_ptr<remote_segment> segment;
-        /// Batch readers that can be used to scan the segment
-        std::list<std::unique_ptr<remote_segment_batch_reader>> readers;
-        /// Reader access time
-        ss::lowres_clock::time_point atime;
-        /// List hook for the list of all materalized segments
-        intrusive_list_hook _hook;
-    };
+    friend struct materialized_segment_state;
 
     using materialized_segment_ptr
       = std::unique_ptr<materialized_segment_state>;
@@ -344,7 +292,6 @@ private:
     iterator seek_by_timestamp(model::timestamp);
 
     using segment_map_t = absl::btree_map<kafka::offset, segment_state>;
-
     using evicted_resource_t = std::variant<
       std::unique_ptr<remote_segment_batch_reader>,
       ss::lw_shared_ptr<remote_segment>>;
@@ -370,6 +317,7 @@ private:
       &materialized_segment_state::_hook>
       _materialized;
     ss::condition_variable _cvar;
+
     /// Timer use to periodically evict stale readers
     ss::timer<ss::lowres_clock> _stm_timer;
     simple_time_jitter<ss::lowres_clock> _stm_jitter;
